Rewrite determinant function to eliminate unnecessary iteration
A [1:] slice is a type of iteration (therefore O(n)) time, so that was removed in favor of offsetting the calculations within the function. Additionally, instead of removing the column with a list comprehension and then converting it to a list, by removing the check for the length of the matrix row, the conversion to a list is unnecesary so it can be left as a generator function (which adds no time complexity).
Finally, I added another shortcut case for a 3x3 matrix as detailed [on Wikipedia](https://en.wikipedia.org/wiki/Determinant#Sarrus'_scheme).
I really encourage you to take the time to understand what this actually does. I personally find it incredibly interesting.